+++ src/core/constants.js
@@ -1,7 +1,7 @@
 const CONSTANTS = {
-    NEXO_VERSION: "2.2.3-1",
+    NEXO_VERSION: "2.2.3-2",
     NEXO_CODE: "Alpha Version:2+2+3:UNQ+COM+BOTHCORE",
-    NODE_CORE_VERSION: process.versions.code,
+    NODE_CORE_VERSION: process.versions.node,
     NODE_V8_VERSION: process.versions.v8,
     NODE_CPP_VERSION: process.versions.modules
 }