+++ src/modules/search.js
@@ -14,16 +14,12 @@ function generateSearchStrings(){
         for (let article of db.site.posts) {
             let href = db.file(article.path('website'));
             let atitle = article.title;
-            let acontent;
+            let acontent='';
             if(title)acontent+='%%%'+article.title;
             if(id)acontent+='%%%'+article.id;
-            if(content)acontent+='%%%'+article.content.replace(/\n/g,' ').replace(/[#*]/g,'');
+            if(content)acontent+='%%%'+article.content.replace(/(\n)|([\u0020\n]{2,})/g,' ').replace(/[#*]/g,'');
             if(date)acontent+='%%%'+article.date.toDateString();
             arr.push({'content':acontent,href,atitle});
-            //if (title) { arr.push({ 'content': article.title, href, atitle }); }
-            //if (id) { arr.push({ 'content': article.id, href, atitle }); }
-            //if (content) { arr.push({ 'content': article.content.replace(/\n/g, ' '), href, atitle }); }
-            //if (date) { arr.push({ 'content': article.date.toDateString(), href, atitle }); }
         }
         writeFile(join(db.dirs.public, 'searchStrings.json'), JSON.stringify(arr), () => { });
     }