+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # CHANGELOG
 
-## 2.2.3-2 [NOW,EDITING,DEVEL] - 2024-03-04
+## 2.2.3-2 - 2024-03-04
 
 ### New
 
 * build mode for theme to better organize (release/devel)
+
+### Fixed
+
+* whitespace in (parsed) post title
+* `undefined` and too much whitespace in searchStrings.json
 
 ## 2.2.3-1 - 2024-03-04